Some fun stuff
Implemented "City Walls" aura for Civic Centers. Test the Romans to see
the Wall ring. I would actually like the ring to show up only when you
have selected walls to build them instead of being a persistent CC decal
as it is now.

Added a cool  aura ring texture to the Roman Temple of Vesta as a test
to see how aura rings could look.

Changed the costs and effects of the Levy techs.